server/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.31.0
python-multipart==0.0.5
python-dotenv==1.0.0
jinja2==3.1.2
numpy==2.2.4